refactor: Modal into OOP
