+++ src/common/count/clickCount.js
@@ -1,7 +1,9 @@
+//TODO: object Clicker with array of numbers : each time click on icon, use the next number before displaying popup
+
 export function clickCountInit() {
   chrome.runtime.onInstalled.addListener(function (details) {
     if (details.reason === "install" || details.reason === "update") {
-      chrome.storage.sync.set({"clickIconCount": 49});
+      chrome.storage.sync.set({"clickIconCount": 20});
     }
   });
 }
@@ -12,6 +14,6 @@ export function updateClickIconCount() {
   });
 }
 
-export function resetClickCount(count) {
+export function resetClickCount(count = 50) {
   chrome.storage.sync.set({"clickIconCount": count}); //100
 }
