+++ src/tab/triggers.js
@@ -1,13 +1,13 @@
 import {detectPageLoad, domainExportChecker, domainLoadChecker} from "./checker/domainChecker";
 import {launchExport} from "./exportProcessor/exportProcess";
 import {uiEnhancer} from "./uiEnhancer/uiEnhancer";
-import {appendModalDetectClicks} from "./uiEnhancer/modals/appendModals";
+import ModalDetectClicks from "./uiEnhancer/modals/ModalDetectClicks";
 
 export async function actionExtensionIconClicked() {
   const domainPage = await domainExportChecker();
   if (domainPage === null) return;
   launchExport(domainPage);
-  appendModalDetectClicks();
+  new ModalDetectClicks(domainPage).appendModal();
 }
 
 export async function actionPageLoaded() {
